code/requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
pathway